Move go/client to go/runtime/client
This makes it more obvious what kind of a client it is as we have a
bunch of different clients and this one is for interacting with
runtimes.
